requirements.txt
streamlit
requests
beautifulsoup4
opencv-python
rembg
fpdf
openai>=1.0.0
python-dotenv
Pillow
numpy
torch>=2.0.0
git+https://github.com/facebookresearch/segment-anything.git
